@@ -1,8 +1,9 @@
 {
-  "name": "kingoverroad",
+  "name": "evofuse2048",
   "private": true,
-  "version": "0.0.0",
+  "version": "1.1.0",
   "main": "electron/main.js",
+  "type": "module",
   "description": "King Over Road",
   "author": "King Over Road",
   "scripts": {
@@ -18,12 +19,14 @@
     "eject": "react-scripts eject",
     "electron-dev": "concurrently \"npm run dev\" \"electron .\"",
     "electron": "electron .",
-    "electron-build": "npm run build && electron-builder"
+    "electron-build": "tsc -b && vite build && electron-builder"
   },
   "dependencies": {
+    "@aws-sdk/client-s3": "^3.864.0",
     "@clerk/clerk-react": "^5.24.1",
     "@fontsource/cinzel-decorative": "^5.2.5",
     "@fontsource/patrick-hand": "^5.2.5",
+    "adm-zip": "^0.5.16",
     "axios": "^1.8.3",
     "bip39": "^3.1.0",
     "electron-is-dev": "^3.0.1",
@@ -37,6 +40,7 @@
     "react-image-crop": "^11.0.10",
     "react-router-dom": "^7.2.0",
     "react-toastify": "^11.0.5",
+    "request": "^2.88.2",
     "styled-components": "^6.1.15",
     "styled-normalize": "^8.1.1",
     "tree-kill": "^1.2.2",
@@ -52,8 +56,8 @@
     "@vitejs/plugin-react": "^4.3.4",
     "autoprefixer": "10.4.18",
     "concurrently": "^9.1.2",
-    "electron": "^35.0.1",
-    "electron-builder": "^25.1.8",
+    "electron": "^37.3.0",
+    "electron-builder": "^26.0.12",
     "eslint": "^9.21.0",
     "eslint-plugin-react-hooks": "^5.1.0",
     "eslint-plugin-react-refresh": "^0.4.19",
@@ -68,10 +72,10 @@
     "vite": "^6.2.0"
   },
   "build": {
-    "appId": "com.kingoverroad.app",
-    "productName": "King Over Road",
+    "appId": "com.evofuse2048.app",
+    "productName": "EvoFuse2048",
     "files": [
-      "build/**/*",
+      "dist/**/*",
       "electron/**/*"
     ],
     "extraResources": [
