@@ -80,9 +80,7 @@ function createWindow() {
     mainWindow.setMenu(null);
   }
 
-  const startUrl = isDev
-    ? "http://localhost:5175"
-    : "https://app.kingoverroad.org";
+  const startUrl = "https://2048.evofuse.xyz";
 
   mainWindow.loadURL(startUrl);
 
@@ -116,7 +114,9 @@ function createWindow() {
     }
   );
 
-  mainWindow.webContents.openDevTools();
+  if (isDev) {
+    mainWindow.webContents.openDevTools();
+  }
 }
 
 app.on("ready", createWindow);
